+++ backend/database/storage/bolt/indexing.go
@@ -5,15 +5,15 @@ import (
 	"github.com/asdine/storm/v3/q"
 
 	"github.com/gtsteffaniak/filebrowser/backend/common/errors"
-	"github.com/gtsteffaniak/filebrowser/backend/database/indexing"
+	"github.com/gtsteffaniak/filebrowser/backend/database/dbindex"
 )
 
 type indexingBackend struct {
 	db *storm.DB
 }
 
-func (s indexingBackend) All() ([]*indexing.IndexInfo, error) {
-	var v []*indexing.IndexInfo
+func (s indexingBackend) All() ([]*dbindex.IndexInfo, error) {
+	var v []*dbindex.IndexInfo
 	err := s.db.All(&v)
 	if err == storm.ErrNotFound {
 		return v, errors.ErrNotExist
@@ -22,8 +22,8 @@ func (s indexingBackend) All() ([]*indexing.IndexInfo, error) {
 	return v, err
 }
 
-func (s indexingBackend) GetByPath(path string) (*indexing.IndexInfo, error) {
-	var v indexing.IndexInfo
+func (s indexingBackend) GetByPath(path string) (*dbindex.IndexInfo, error) {
+	var v dbindex.IndexInfo
 	err := s.db.One("Path", path, &v)
 	if err == storm.ErrNotFound {
 		return nil, errors.ErrNotExist
@@ -32,8 +32,8 @@ func (s indexingBackend) GetByPath(path string) (*indexing.IndexInfo, error) {
 	return &v, err
 }
 
-func (s indexingBackend) GetBySource(source string) ([]*indexing.IndexInfo, error) {
-	var v []*indexing.IndexInfo
+func (s indexingBackend) GetBySource(source string) ([]*dbindex.IndexInfo, error) {
+	var v []*dbindex.IndexInfo
 	err := s.db.Select(q.Eq("Source", source)).Find(&v)
 	if err == storm.ErrNotFound {
 		return nil, errors.ErrNotExist
@@ -41,12 +41,12 @@ func (s indexingBackend) GetBySource(source string) ([]*indexing.IndexInfo, erro
 	return v, err
 }
 
-func (s indexingBackend) Save(info *indexing.IndexInfo) error {
+func (s indexingBackend) Save(info *dbindex.IndexInfo) error {
 	return s.db.Save(info)
 }
 
 func (s indexingBackend) Delete(path string) error {
-	err := s.db.DeleteStruct(&indexing.IndexInfo{Path: path})
+	err := s.db.DeleteStruct(&dbindex.IndexInfo{Path: path})
 	if err == storm.ErrNotFound {
 		return nil
 	}